Slightly more intuitive saving UI, tries to fix #240.
On save, move focus to the commit message dialog, to encouage users to enter message.

In order to make this process more fluid, the default commit message was moved to `placeholder`, and is swapped in for `val()` on save if none is present.

May make sense to have someone with more design/UI sense take a pass, perhaps the commit box border changes on focus, the commit button changes color, etc. to stand out?